# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VatanSoft Case",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "summary": "Get token for login",
                "operationId": "8dcb70df1020986038d098cc08d05dae",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User email",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "email"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An example resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "description": "access_token",
                                            "type": "string",
                                            "default": ""
                                        },
                                        "token_type": {
                                            "description": "token_type",
                                            "type": "string",
                                            "default": "bearer"
                                        },
                                        "expires_in": {
                                            "description": "expires_in",
                                            "type": "integer",
                                            "default": "3060"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/me": {
            "get": {
                "summary": "Get logged in user information",
                "operationId": "b9abb1a7a74670a19c215c2c133f14d8",
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/auth/logout": {
            "post": {
                "summary": "Logout from current user",
                "operationId": "69281b12abb272c76871f19cb17ca563",
                "responses": {
                    "200": {
                        "description": "An example resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "message",
                                            "type": "string",
                                            "default": "Successfully logged out"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/messages": {
            "post": {
                "summary": "Send a new message",
                "operationId": "db79bb7bcadfdb708e9f6c38ac928f4a",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Message title",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "content",
                        "in": "query",
                        "description": "Message Content",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An example resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "status",
                                            "type": "boolean",
                                            "default": "success"
                                        },
                                        "message": {
                                            "description": "message",
                                            "type": "string",
                                            "default": "SMS gönderim talebi işleme alındı."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/message-reports": {
            "get": {
                "summary": "Get Message Reports",
                "description": "Display a listing of the resource.",
                "operationId": "4c7932946ae52262cdebdd9b84ecd90b",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "description": "date of reports",
                        "required": false,
                        "schema": {
                            "type": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An example resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "data",
                                            "type": "array",
                                            "items": {
                                                "type": "array",
                                                "items": {}
                                            }
                                        },
                                        "status": {
                                            "description": "status",
                                            "type": "string",
                                            "default": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/message-reports/{number}": {
            "get": {
                "summary": "Get detail of report",
                "description": "Display the specified resource.",
                "operationId": "41b23f769482d81a31d5169bb0f04d22",
                "responses": {
                    "200": {
                        "description": "An example resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "data",
                                            "type": "array",
                                            "items": {
                                                "type": "array",
                                                "items": {}
                                            }
                                        },
                                        "status": {
                                            "description": "status",
                                            "type": "string",
                                            "default": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}